+++ apps/create-plugin/template/src/Project.ts.t
@@ -1,12 +1,15 @@
 import {
   type TransactionFilter,
-  type SwapActionParams,
+    {{#each actionTypes}}
+  type {{capitalize this}}ActionParams,
+  {{/each}}
   compressJson,
 } from '@rabbitholegg/questdk'
 import { type Address } from 'viem'
+import { Chains } from '@rabbitholegg/questdk-plugin-utils'
 
 /*
- * Function template for handling various blockchain action types.
+ * Function templates for handling various blockchain action types.
  * It's adaptable for actions defined in ActionParams: Bridge, Swap, Stake, Mint, Delegate, Quest, Etc.
  * Duplicate and customize for each specific action type.
  * If you wish to use a different action other than swap, import one of the ActionParams types
@@ -14,11 +17,16 @@ import { type Address } from 'viem'
  * the action params you wish to use.
  */
 
-export const swap = async (
-  swap: SwapActionParams,
-): Promise<TransactionFilter> => {
-  // This is the information we'll use to compose the Transaction object
-  const { chainId, contractAddress, tokenIn } = swap
+  {{#each actionTypes}}
+  export const {{lowercase this}} = async(_params: {{capitalize this}}ActionParams): Promise<TransactionFilter> => {
+
+    // the ActionParams for this function are populated in the Boost Manager when the actual Boost is launched.
+    
+    // In this function you should load the ABI, and translate any ActionParams into the input object defined below 
+    // which should match the parameter names in the transaction
+
+    // You can also use the boostdk filter system to support operators on parameters, for example, greater than
+
 
   // We always want to return a compressed JSON object which we'll transform into a TransactionFilter
   return compressJson({
@@ -26,7 +34,11 @@ export const swap = async (
     to: '0x0', // The to field is the address of the contract we're interacting with
     input: {}, // The input object is where we'll put the ABI and the parameters
   })
-}
+
+  }
+
+  {{/each}}
+
 
 export const getSupportedTokenAddresses = async (
   _chainId: number,
@@ -37,5 +49,5 @@ export const getSupportedTokenAddresses = async (
 
 export const getSupportedChainIds = async (): Promise<number[]> => {
   // This should return all of the ChainIds that are supported by the Project we're integrating
-  return []
+  return [{{#each chains}}Chains.{{this}}, {{/each}}]
 }